requirements.txt
flask
psycopg2-binary
gunicorn
requests
beautifulsoup4
lxml
edgartools
yfinance
